cleanup

+++ commons.h
@@ -64,13 +64,15 @@ char* shift_args(int *argc, char ***argv)
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/stat.h>
+#include <stdint.h>
 
-unsigned long long file_size(const char* file_path){
+uint64_t file_size(const char* file_path)
+{
     struct stat file;
     if (stat(file_path, &file) == -1){
         return 0;
     }
-    return (unsigned long long) file.st_size;
+    return (uint64_t) file.st_size;
 }
 
 // allocate and populate a string with the file's content
@@ -79,7 +81,7 @@ char* read_entire_file(char *file_path)
     if (file_path == NULL) return NULL;
     FILE *file = fopen(file_path, "r");
     if (file == NULL) return NULL;
-    unsigned long long size = file_size(file_path);
+    uint64_t size = file_size(file_path);
     char *content = (char*) calloc(size+1, sizeof(*content));
     if (!content){
         fclose(file);
